Extend config to support MapeHeat
ref #4

NuHeat is now branded as MapeHeat and new thermostats are sold under the
new brand, however old thermostats appear to continue to be supported
with their original NuHeat brand. There is a new URL that uses the same
API for MapeHeat products and so this library should support both, with
NuHeat remaining as the default.